+++ android/src/main/java/com/twiliovoicereactnative/NotificationUtility.java
@@ -1,6 +1,7 @@
 package com.twiliovoicereactnative;
 
 import java.security.SecureRandom;
+
 import java.util.Map;
 import java.util.Objects;
 
@@ -32,42 +33,93 @@ import com.twiliovoicereactnative.CallRecordDatabase.CallRecord;
 
 class NotificationUtility {
   private static final SecureRandom secureRandom = new SecureRandom();
+
   private static class NotificationResource {
+    enum Type { INCOMING, OUTGOING, ANSWERED }
+
+    private final Type type;
     private final Context ctx;
-    private final String smallIconId;
-    private final String contentTextId;
-    public NotificationResource(Context context, String smallIconId, String contentTextId) {
-      ctx = context;
-      this.smallIconId = smallIconId;
-      this.contentTextId = contentTextId;
+    private final CallRecord callRecord;
+
+    public NotificationResource(Context context, Type type, CallRecord callRecord) {
+      this.ctx = context;
+      this.type = type;
+      this.callRecord = callRecord;
     }
-    int getSmallIconId() {
-      return getDrawableResourceId(smallIconId);
+
+    private int getSmallIconId() {
+      switch (this.type) {
+        case INCOMING:
+          return this.getDrawableResourceId("incoming_call_small_icon");
+        case OUTGOING:
+          return this.getDrawableResourceId("outgoing_call_small_icon");
+        default:
+          return this.getDrawableResourceId("answered_call_small_icon");
+      }
     }
-    public String getContent(CallInvite callInvite) {
-      final String from = (null != callInvite) ? getDisplayName(callInvite) : "";
-      return ctx.getString(getTextResourceId(contentTextId))
-        .replaceAll("\\$\\{from\\}", from);
+
+    private String getContentText() {
+      switch (this.type) {
+        case INCOMING:
+          return this.ctx.getString(this.getTextResourceId("incoming_call_caller_name_text"));
+        case OUTGOING:
+          return this.ctx.getString(this.getTextResourceId("outgoing_call_caller_name_text"));
+        default:
+          return this.ctx.getString(this.getTextResourceId("answered_call_caller_name_text"));
+      }
     }
-    public String getContent(String to) {
-      return ctx.getString(getTextResourceId(contentTextId))
-        .replaceAll("\\$\\{to\\}", to);
+
+    public String getName() {
+      if (this.callRecord.getDirection() == CallRecord.Direction.INCOMING) {
+        final String template = ConfigurationProperties.getIncomingCallContactHandleTemplate();
+        if (template != null) {
+          return templateDisplayName(template, this.callRecord.getCustomParameters());
+        }
+
+        final CallInvite callInvite = this.callRecord.getCallInvite();
+        final String from = null != callInvite ? getDisplayName(callInvite) : "";
+        return this.getContentText().replaceAll("\\$\\{from\\}", from);
+      }
+
+      // this.callRecord.Direction == CallRecord.Direction.OUTGOING
+      if (this.callRecord.getNotificationDisplayName() != null) {
+        return this.callRecord.getNotificationDisplayName();
+      }
+
+      final String to = this.callRecord.getCallRecipient();
+      return this.getContentText().replaceAll("\\$\\{to\\}", to);
     }
 
     @SuppressLint("DiscouragedApi")
     private int getDrawableResourceId(String id) {
       return ctx.getResources().getIdentifier(id, "drawable", ctx.getPackageName());
     }
+
     @SuppressLint("DiscouragedApi")
     private int getTextResourceId(String id) {
       return ctx.getResources().getIdentifier(id, "string", ctx.getPackageName());
     }
+
     private static String getDisplayName(@NonNull CallInvite callInvite) {
       final String title = callInvite.getFrom();
       if (title.startsWith("client:")) {
         return title.replaceFirst("client:", "");
       }
       return title;
+    }
+
+    private static String templateDisplayName(final String template, final Map<String, String> twimlParams) {
+      String processedTemplate = template;
+
+      for (Map.Entry<String, String> e : twimlParams.entrySet()) {
+        String paramKey = e.getKey();
+        String paramValue = e.getValue();
+        processedTemplate = processedTemplate.replaceAll(
+          String.format("\\$\\{%s\\}", paramKey),
+          paramValue);
+      }
+
+      return processedTemplate;
     }
   }
 
@@ -80,11 +132,11 @@ class NotificationUtility {
                                                             @NonNull final String channelImportance) {
     final NotificationResource notificationResource = new NotificationResource(
       context,
-      "incoming_call_small_icon",
-      "incoming_call_caller_name_text");
+      NotificationResource.Type.INCOMING,
+      callRecord);
 
     final Person incomingCaller = new Person.Builder()
-      .setName(notificationResource.getContent(callRecord.getCallInvite()))
+      .setName(notificationResource.getName())
       .build();
 
     Intent foregroundIntent = constructMessage(
@@ -124,11 +176,11 @@ class NotificationUtility {
                                                                              @NonNull final CallRecord callRecord) {
     final NotificationResource notificationResource = new NotificationResource(
       context,
-      "answered_call_small_icon",
-      "answered_call_caller_name_text");
+      NotificationResource.Type.ANSWERED,
+      callRecord);
 
     final Person activeCaller = new Person.Builder()
-      .setName(notificationResource.getContent(callRecord.getCallInvite()))
+      .setName(notificationResource.getName())
       .build();
 
     Intent foregroundIntent = constructMessage(
@@ -161,11 +213,11 @@ class NotificationUtility {
                                                                              @NonNull final CallRecord callRecord) {
     final NotificationResource notificationResource = new NotificationResource(
       context,
-      "outgoing_call_small_icon",
-      "outgoing_call_caller_name_text");
+      NotificationResource.Type.OUTGOING,
+      callRecord);
 
     final Person activeCaller = new Person.Builder()
-      .setName(notificationResource.getContent(callRecord.getCallRecipient()))
+      .setName(notificationResource.getName())
       .build();
 
     Intent foregroundIntent = constructMessage(