+++ src/__tests__/Voice.test.ts
@@ -794,6 +794,35 @@ describe('Voice class', () => {
         ).resolves.toBeUndefined();
       });
     });
+
+    describe('.setIncomingCallContactHandleTemplate', () => {
+      it('invokes the native module with a string', async () => {
+        const template = 'Foo ${DisplayName}';
+        await new Voice().setIncomingCallContactHandleTemplate(template);
+        expect(
+          jest.mocked(
+            MockNativeModule.voice_setIncomingCallContactHandleTemplate
+          ).mock.calls
+        ).toEqual([[template]]);
+      });
+
+      it('invokes the native module with no parameter', async () => {
+        await new Voice().setIncomingCallContactHandleTemplate();
+        expect(
+          jest.mocked(
+            MockNativeModule.voice_setIncomingCallContactHandleTemplate
+          ).mock.calls
+        ).toEqual([[]]);
+      });
+
+      it('returns a Promise<void>', async () => {
+        const setIncomingCallContactHandleTemplatePromise =
+          new Voice().setIncomingCallContactHandleTemplate('foobar');
+        await expect(
+          setIncomingCallContactHandleTemplatePromise
+        ).resolves.toBeUndefined();
+      });
+    });
   });
 
   describe('private methods', () => {