+++ src/index.tsx
@@ -9,9 +9,7 @@
  *
  * @packageDocumentation
  */
-
 export { Voice } from './Voice';
-
 export { AudioDevice } from './AudioDevice';
 export { Call } from './Call';
 export { CallInvite } from './CallInvite';