+++ src/type/NativeModule.ts
@@ -61,7 +61,8 @@ export interface TwilioVoiceReactNative extends NativeModulesStatic {
    */
   voice_connect_android(
     token: string,
-    twimlParams: Record<string, any>
+    twimlParams: Record<string, any>,
+    notificationDisplayName: string | undefined
   ): Promise<NativeCallInfo>;
   voice_connect_ios(
     token: string,
@@ -72,6 +73,7 @@ export interface TwilioVoiceReactNative extends NativeModulesStatic {
   voice_setCallKitConfiguration(
     configuration: Record<string, any>
   ): Promise<void>;
+  voice_setIncomingCallContactHandleTemplate(template?: string): Promise<void>;
   voice_getAudioDevices(): Promise<NativeAudioDevicesInfo>;
   voice_getCalls(): Promise<NativeCallInfo[]>;
   voice_getCallInvites(): Promise<NativeCallInviteInfo[]>;